requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
alembic
antlr4-python3-runtime
asteroid-filterbanks
asttokens
attrs
audioread
cachetools
certifi
cffi
charset-normalizer
click
colorlog
comm
contourpy
cycler
debugpy
decorator
docopt
einops
executing
ffmpeg
filelock
fonttools
frozenlist
fsspec
google-api-core
google-auth
google-cloud-core
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
greenlet
huggingface-hub
HyperPyYAML
idna
imageio
imageio-ffmpeg
ipykernel
ipython
jedi
Jinja2
joblib
julius
jupyter_client
jupyter_core
kiwisolver
lazy_loader
librosa
lightning
lightning-utilities
llvmlite
Mako
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
more-itertools
moviepy
mpmath
msgpack
multidict
nest-asyncio
networkx
numba
numpygit checkout frontend-nextjs -- applications
omegaconf
openai-whisper
opencv-python
optuna
packaging
pandas
parso
pexpect
pillow
platformdirs
pooch
primePy
proglog
prompt_toolkit
propcache
proto-plus
protobuf
psutil
ptyprocess
pure_eval
pyannote.audio
pyannote.core
pyannote.database
pyannote.metrics
pyannote.pipeline
pyasn1
pyasn1_modules
pycparser
pydub
Pygments
pyparsing
python-dateutil
python-dotenv
pytorch-lightning
pytorch-metric-learning
pytz
PyYAML
pyzmq
regex
requests
rich
rsa
ruamel.yaml
ruamel.yaml.clib
safetensors
scikit-learn
scipy
semver
sentencepiece
setuptools
shellingham
six
sortedcontainers
soundfile
soxr
speechbrain
SQLAlchemy
stack-data
sympy
tabulate
tensorboardX
threadpoolctl
tiktoken
tokenizers
torch
torch-audiomentations
torch_pitch_shift
torchaudio
torchmetrics
tornado
tqdm
traitlets
transformers
triton
typer
typing_extensions
tzdata
urllib3
wcwidth
whisper
yarl
